+++ packages/plugin-types/src/index.ts
@@ -253,4 +253,4 @@ export { createConsoleLogger, createNoopLogger } from './common/logger';
  * Version of the plugin-types package.
  * Can be used at runtime to check compatibility.
  */
-export const PLUGIN_TYPES_VERSION = '2.6.0';
+export const PLUGIN_TYPES_VERSION = '2.7.0';